+++ tests/Harvest/Importers/ItemImporterTest.php
@@ -2,11 +2,15 @@
 
 namespace Tests\Import\Importers;
 
+use App\Authority;
 use App\Harvest\Importers\ItemImporter;
+use App\Harvest\Mappers\AuthorityItemMapper;
 use App\Harvest\Mappers\ItemImageMapper;
-use App\Harvest\Mappers\CollectionMapper;
+use App\Harvest\Mappers\CollectionItemMapper;
 use App\Harvest\Mappers\ItemMapper;
 use App\Harvest\Result;
+use App\Item;
+use App\ItemImage;
 use Illuminate\Foundation\Testing\DatabaseMigrations;
 use Tests\TestCase;
 
@@ -20,12 +24,47 @@ class ItemImporterTest extends TestCase
         $importer->import($row, $result = new Result());
     }
 
-    public function testUpdatedRelations() {
+    public function testDeleteRelations() {
         $row = $this->getData();
         $importer = $this->initImporter($row);
+        $item = factory(Item::class)->create(['id' => 'SVK:SNG.G_10044']);
+        $image = factory(ItemImage::class)->make(['iipimg_url' => 'to_be_deleted']);
+        $item->images()->save($image);
+
+        $item->load('images');
+        $this->assertTrue($item->images->contains(function ($i, ItemImage $image) {
+            return $image->iipimg_url === 'to_be_deleted';
+        }));
+
+        $item->load('images');
+        $item = $importer->import($row, $result = new Result());
+
+        $this->assertCount(2, $item->images);
+        $this->assertFalse($item->images->contains(function ($i, ItemImage $image) {
+            return $image->iipimg_url === 'to_be_deleted';
+        }));
+    }
+
+    public function testDetachRelations() {
+        $row = $this->getData();
+        $importer = $this->initImporter($row);
+        $item = factory(Item::class)->create(['id' => 'SVK:SNG.G_10044']);
+        factory(Authority::class)->create(['id' => 1922]);
+        factory(Authority::class)->create(['id' => 10816]);
+        $authority = factory(Authority::class)->create(['id' => 'to_be_detached']);
+        $item->authorities()->attach($authority);
+
+        $item->load('authorities');
+        $this->assertTrue($item->authorities->contains(function ($i, Authority $authority) {
+            return $authority->id === 'to_be_detached';
+        }));
 
         $item = $importer->import($row, $result = new Result());
-        $this->assertCount(2, $item->images);
+
+        $this->assertCount(2, $item->authorities);
+        $this->assertFalse($item->authorities->contains(function ($i, Authority $authority) {
+            return $authority->id === 'to_be_detached';
+        }));
     }
 
     protected function getData() {
@@ -103,7 +142,14 @@ class ItemImporterTest extends TestCase
                 'urn:svk:psi:per:sng:0000010816',
                 'Teniers, David',
             ],
-            'creator_role' => [],
+            'creator_role' => [
+                'autor/author',
+                'autor/author',
+            ],
+            'authorities' => [
+                ['id' => ['urn:svk:psi:per:sng:0000001922']],
+                ['id' => ['urn:svk:psi:per:sng:0000010816']],
+            ],
             'rights' => [
                 '1',
                 'publikovať/public',
@@ -135,7 +181,8 @@ class ItemImporterTest extends TestCase
         $importer = new ItemImporter(
             $itemMapperMock = $this->getMock(ItemMapper::class),
             $itemImageMapperMock = $this->getMock(ItemImageMapper::class),
-            $collectionMapper = $this->getMock(CollectionMapper::class)
+            $collectionItemMapperMock = $this->getMock(CollectionItemMapper::class),
+            $authorityItemMapperMock = $this->getMock(AuthorityItemMapper::class)
         );
         $itemMapperMock
             ->expects($this->once())
@@ -197,6 +244,12 @@ class ItemImporterTest extends TestCase
                 'work_level:cs' => null,
             ])
         ;
+        $itemMapperMock
+            ->expects($this->once())
+            ->method('mapId')
+            ->with($row)
+            ->willReturn('SVK:SNG.G_10044');
+        ;
         $itemImageMapperMock
             ->expects($this->exactly(2))
             ->method('map')
@@ -213,6 +266,17 @@ class ItemImporterTest extends TestCase
                 ]
             )
         ;
+        $authorityItemMapperMock
+            ->expects($this->exactly(2))
+            ->method('map')
+            ->withConsecutive(
+                [$row['authorities'][0]],
+                [$row['authorities'][1]]
+            )
+            ->willReturnOnConsecutiveCalls(
+                ['id' => 1922],
+                ['id' => 10816]
+            );
 
         return $importer;
     }